requirements.txt
fastapi[standard]==0.116.1
sqlmodel==0.0.24
python-dotenv==1.1.1
google-cloud-texttospeech==2.29.0
pydantic-settings==2.10.1